+++ ios/CDVBackgroundGeolocation/CDVBackgroundGeolocation.m
@@ -202,7 +202,7 @@ static NSString * const TAG = @"CDVBackgroundGeolocation";
 - (void) getPluginVersion:(CDVInvokedUrlCommand*)command
 {
     NSLog(@"%@ #%@", TAG, @"getPluginVersion");
-    NSString *version = @"3.0.0"; // keep in sync with plugin.xml
+    NSString *version = @"3.2.0"; // keep in sync with plugin.xml
     CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:version];
     [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
 }
@@ -352,6 +352,46 @@ static NSString * const TAG = @"CDVBackgroundGeolocation";
 {
     [self.commandDelegate runInBackground:^{
         NSInteger count = [facade getPendingSyncCount];
+        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsNSInteger:count];
+        [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+    }];
+}
+
+- (void) startSession:(CDVInvokedUrlCommand*)command
+{
+    [self.commandDelegate runInBackground:^{
+        [facade startSession];
+        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+        [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+    }];
+}
+
+- (void) getSessionLocations:(CDVInvokedUrlCommand*)command
+{
+    [self.commandDelegate runInBackground:^{
+        NSArray *locations = [facade getSessionLocations];
+        NSMutableArray *dictionaryLocations = [[NSMutableArray alloc] initWithCapacity:[locations count]];
+        for (MAURLocation *location in locations) {
+            [dictionaryLocations addObject:[location toDictionaryWithId]];
+        }
+        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsArray:dictionaryLocations];
+        [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+    }];
+}
+
+- (void) clearSession:(CDVInvokedUrlCommand*)command
+{
+    [self.commandDelegate runInBackground:^{
+        [facade clearSession];
+        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+        [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+    }];
+}
+
+- (void) getSessionLocationsCount:(CDVInvokedUrlCommand*)command
+{
+    [self.commandDelegate runInBackground:^{
+        NSInteger count = [facade getSessionLocationsCount];
         CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsNSInteger:count];
         [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
     }];